+++ src/functions/Chocolatey-Pack.ps1
@@ -4,21 +4,22 @@ param(
 )
   Write-Debug "Running 'Chocolatey-Pack' for $packageName. If nuspec name is not passed, it will find the nuspec file in the current working directory";
 
-  $packageArgs = "pack $packageName -NoPackageAnalysis"
+  $packageArgs = "pack $packageName -NoPackageAnalysis -NonInteractive"
   $logFile = Join-Path $nugetChocolateyPath 'pack.log'
   $errorLogFile = Join-Path $nugetChocolateyPath 'error.log'
-  
+
   Write-Host "Calling `'$nugetExe $packageArgs`'."
-  
-  Start-Process $nugetExe -ArgumentList $packageArgs -NoNewWindow -Wait -RedirectStandardOutput $logFile -RedirectStandardError $errorLogFile
+
+  $process = Start-Process $nugetExe -ArgumentList $packageArgs -NoNewWindow -Wait -RedirectStandardOutput $logFile -RedirectStandardError $errorLogFile -PassThru
+  # this is here for specific cases in Posh v3 where -Wait is not honored
+  try { if (!($process.HasExited)) { Wait-Process $process } } catch { }
 
   $nugetOutput = Get-Content $logFile -Encoding Ascii
   foreach ($line in $nugetOutput) {
     Write-Host $line
   }
   $errors = Get-Content $errorLogFile
-  if ($errors -ne '') {
-    Write-Host $errors -BackgroundColor Red -ForegroundColor White
-    #throw $errors
+  if ($process.ExitCode -ne 0) {
+    throw $errors
   }
 }